+++ drop-down-terminal@gs-extensions.zzrough.org/convenience.js
@@ -21,10 +21,16 @@ const MainLoop = imports.mainloop;
 const Gdk = imports.gi.Gdk;
 const Gio = imports.gi.Gio;
 const GLib = imports.gi.GLib;
+const Gtk = imports.gi.Gtk;
 
-const GLIB_VERSION = GLib.MAJOR_VERSION * 1000
-                   + GLib.MINOR_VERSION * 10
+
+const GLIB_VERSION = GLib.MAJOR_VERSION * 10000
+                   + GLib.MINOR_VERSION * 100
                    + GLib.MICRO_VERSION;
+
+const GTK_VERSION = Gtk.MAJOR_VERSION * 10000
+                  + Gtk.MINOR_VERSION * 100
+                  + Gtk.MICRO_VERSION;
 
 
 function getSettings(extensionPath, extensionId) {
@@ -100,7 +106,7 @@ function gdkRunner(func) {
  * @return the pid
  */
 function getPid() {
-    if (GLIB_VERSION >= 2354) { // only available in GLib 2.35.4+
+    if (GLIB_VERSION >= 23540) { // only available in GLib 2.35.4+
         return new Gio.Credentials().get_unix_pid();
     } else {
         return imports.system.getpid();