+++ NEWS.md
@@ -59,3 +59,4 @@ v9 (2013-03-25)
 v10 (2013-04-21)
 - bugfix: [Allow a block or underline cursor in the terminal (github #36)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/36)
 - bugfix: [Switch focus faster (github #38)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/38)
+- code: get rid of the gtk allocation regression fix for gtk >= 3.8.1